Refactored Swift Storage Service helpers into own test case, moved FileSpy into helper file.
